Convert data fetching components into server side components
