+++ components/data-table.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {
     ColumnDef,
     flexRender,
